Prerelease 4.0

+++ Cave.Media/Interop.cs
@@ -19,7 +19,7 @@ public static class Interop
         public static extern IntPtr memcpy(IntPtr dest, IntPtr src, int count);
 
         [DllImport("msvcrt", EntryPoint = "memcpy", CallingConvention = CallingConvention.Cdecl, SetLastError = false)]
-        public static unsafe extern IntPtr memcpy(int* dest, int* src, int count);
+        public static unsafe extern IntPtr memcpy(byte* dest, byte* src, int count);
 
         [DllImport("msvcrt", EntryPoint = "memset", CallingConvention = CallingConvention.Cdecl, SetLastError = false)]
         public static extern IntPtr memset(IntPtr dest, int c, int count);
